rename to e2core
