+++ bus/bus/hub.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/suborbital/deltav/bus/bus/tunnel"
-	"github.com/suborbital/deltav/bus/bus/withdraw"
+	"github.com/suborbital/e2core/bus/bus/tunnel"
+	"github.com/suborbital/e2core/bus/bus/withdraw"
 	"github.com/suborbital/vektor/vlog"
 )
