+++ e2core/satbackend/orchestrator.go
@@ -13,8 +13,8 @@ import (
 
 	"github.com/suborbital/appspec/appsource"
 	"github.com/suborbital/appspec/appsource/client"
-	"github.com/suborbital/deltav/deltav/satbackend/exec"
-	"github.com/suborbital/deltav/options"
+	"github.com/suborbital/e2core/e2core/satbackend/exec"
+	"github.com/suborbital/e2core/options"
 	"github.com/suborbital/vektor/vlog"
 )
