+++ scheduler/scheduler.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/pkg/errors"
 
-	"github.com/suborbital/deltav/bus/bus"
+	"github.com/suborbital/e2core/bus/bus"
 	"github.com/suborbital/vektor/vlog"
 )
