+++ server/coordinator/coordinator.go
@@ -12,12 +12,12 @@ import (
 	"github.com/suborbital/appspec/fqmn"
 	"github.com/suborbital/appspec/tenant"
 	"github.com/suborbital/appspec/tenant/executable"
-	"github.com/suborbital/deltav/bus/bus"
-	"github.com/suborbital/deltav/bus/transport/kafka"
-	"github.com/suborbital/deltav/bus/transport/nats"
-	"github.com/suborbital/deltav/bus/transport/websocket"
-	"github.com/suborbital/deltav/options"
-	"github.com/suborbital/deltav/server/coordinator/executor"
+	"github.com/suborbital/e2core/bus/bus"
+	"github.com/suborbital/e2core/bus/transport/kafka"
+	"github.com/suborbital/e2core/bus/transport/nats"
+	"github.com/suborbital/e2core/bus/transport/websocket"
+	"github.com/suborbital/e2core/options"
+	"github.com/suborbital/e2core/server/coordinator/executor"
 	"github.com/suborbital/vektor/vk"
 	"github.com/suborbital/vektor/vlog"
 )
