+++ server/coordinator/executor/executor.go
@@ -11,10 +11,10 @@ import (
 	"github.com/suborbital/appspec/capabilities"
 	"github.com/suborbital/appspec/request"
 	"github.com/suborbital/appspec/tenant/executable"
-	"github.com/suborbital/deltav/bus/bus"
-	"github.com/suborbital/deltav/bus/discovery/local"
-	"github.com/suborbital/deltav/bus/transport/websocket"
-	"github.com/suborbital/deltav/scheduler"
+	"github.com/suborbital/e2core/bus/bus"
+	"github.com/suborbital/e2core/bus/discovery/local"
+	"github.com/suborbital/e2core/bus/transport/websocket"
+	"github.com/suborbital/e2core/scheduler"
 	"github.com/suborbital/vektor/vk"
 	"github.com/suborbital/vektor/vlog"
 )
