+++ server/coordinator/sequence/sequence.go
@@ -8,9 +8,9 @@ import (
 
 	"github.com/suborbital/appspec/request"
 	"github.com/suborbital/appspec/tenant/executable"
-	"github.com/suborbital/deltav/bus/bus"
-	"github.com/suborbital/deltav/scheduler"
-	"github.com/suborbital/deltav/server/coordinator/executor"
+	"github.com/suborbital/e2core/bus/bus"
+	"github.com/suborbital/e2core/scheduler"
+	"github.com/suborbital/e2core/server/coordinator/executor"
 	"github.com/suborbital/vektor/vk"
 	"github.com/suborbital/vektor/vlog"
 )
